ccs: move ccs circuit to ir
