+++ src/ccs/backend/mod.rs
@@ -1,2 +1,1 @@
-pub mod ccs;
 pub mod sonobe;
